DesktopTracker/requirements.txt
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.0.0
pystray==0.19.5
pywin32==306
winshell==0.6
pyinstaller==6.3.0
